requirements.txt
discord.py
python-dotenv
feedparser
PyYAML
ruamel.yaml